feat:user refactor
