+++ src/modules/user/infra/db/user.repository.ts
@@ -1,0 +1,6 @@
+import { User } from '../../domain/entity/user.entity';
+import { CustomRepository } from '../../../../shared/typeorm-ex.decorator';
+import { Repository } from 'typeorm';
+
+@CustomRepository({ entity: User })
+export class UserRepository extends Repository<User> {}
