+++ tsconfig.json
@@ -18,5 +18,6 @@
     "strictBindCallApply": false,
     "forceConsistentCasingInFileNames": false,
     "noFallthroughCasesInSwitch": false,
+    "sourceRoot": "/"
   }
 }
